Additional changes for size_t and casting to in for warnings in 32 bit version.
git-svn-id: svn://136.177.114.72/svn_GW/phastpp/trunk@3508 1feff8c3-07ed-0310-ac33-dd36852eb9cd

+++ read.c
@@ -877,7 +877,7 @@ get_line(FILE * fp)
 		}
 		strncpy(line, line_save, (unsigned) j);
 		line[j] = '\0';
-		for (i = 0; i < j; i++)
+		for (i = 0; i < (int) j; i++)
 		{
 			if (!isspace((int) line[i]))
 			{
